Finish view for forecast weather

+++ app/src/main/java/edu/pk/weatherapp/model/WeatherMain.kt
@@ -4,9 +4,9 @@ import com.squareup.moshi.Json
 import edu.pk.weatherapp.R
 
 class WeatherMain(
-        @Json(name = "temp") var temperature: Int,
-        @Json(name = "pressure") var pressure: Int,
-        @Json(name = "temp_min") var tempMin: Int,
-        @Json(name = "temp_max") var tempMax: Int,
-        @Json(name = "humidity") var humidity: Int
+        @Json(name = "temp") var temperature: Double,
+        @Json(name = "pressure") var pressure: Double,
+        @Json(name = "temp_min") var tempMin: Double,
+        @Json(name = "temp_max") var tempMax: Double,
+        @Json(name = "humidity") var humidity: Double
 )
